Move parse config into utils
So we can call it from detection code
